fixed other migration

+++ db/migrate/20140801153034_drop_answers_question_id.rb
@@ -1,6 +1,7 @@
 class DropAnswersQuestionId < ActiveRecord::Migration
   def up
-    remove_column :answers, :question_id
+    remove_column :answers, :question_id if Answer.column_names.include?('question_id')
+    remove_column :answers, :questionable_id if Answer.column_names.include?('questionable_id')
   end
 
   def down
